move open telemetry tests to core tests
