+++ src/NServiceBus.AcceptanceTests/Core/OpenTelemetry/When_incoming_event_has_trace.cs
@@ -1,13 +1,11 @@
-﻿namespace NServiceBus.AcceptanceTests.Diagnostics;
+﻿namespace NServiceBus.AcceptanceTests.Core.OpenTelemetry;
 
 using System;
-using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using AcceptanceTesting;
-using AcceptanceTesting.Customization;
-using EndpointTemplates;
+using NServiceBus.AcceptanceTesting;
+using NServiceBus.AcceptanceTesting.Customization;
+using NServiceBus.AcceptanceTests.EndpointTemplates;
 using NUnit.Framework;
 
 [NonParallelizable] // Ensure only activities for the current test are captured
