+++ src/NServiceBus.AcceptanceTests/Core/OpenTelemetry/When_retrying_messages.cs
@@ -1,10 +1,10 @@
-﻿namespace NServiceBus.AcceptanceTests.Diagnostics;
+﻿namespace NServiceBus.AcceptanceTests.Core.OpenTelemetry;
 
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using AcceptanceTesting;
-using EndpointTemplates;
+using NServiceBus.AcceptanceTesting;
+using NServiceBus.AcceptanceTests.EndpointTemplates;
 using NUnit.Framework;
 
 [NonParallelizable] // Ensure only activities for the current test are captured
