+++ src/NServiceBus.AcceptanceTests/Core/OpenTelemetry/When_sending_replies.cs
@@ -1,8 +1,8 @@
-namespace NServiceBus.AcceptanceTests.Diagnostics
+namespace NServiceBus.AcceptanceTests.Core.OpenTelemetry
 {
     using System.Threading.Tasks;
-    using AcceptanceTesting;
-    using EndpointTemplates;
+    using NServiceBus.AcceptanceTesting;
+    using NServiceBus.AcceptanceTests.EndpointTemplates;
     using NUnit.Framework;
 
     [NonParallelizable] // Ensure only activities for the current test are captured
