+++ src/NServiceBus.AcceptanceTests/Core/OpenTelemetry/When_subscribing.cs
@@ -1,11 +1,11 @@
-﻿namespace NServiceBus.AcceptanceTests.Diagnostics;
+﻿namespace NServiceBus.AcceptanceTests.Core.OpenTelemetry;
 
 using System.Collections.Immutable;
 using System.Linq;
 using System.Threading.Tasks;
-using AcceptanceTesting;
-using EndpointTemplates;
-using Features;
+using NServiceBus.AcceptanceTesting;
+using NServiceBus.AcceptanceTests.EndpointTemplates;
+using NServiceBus.Features;
 using NUnit.Framework;
 
 [NonParallelizable] // Ensure only activities for the current test are captured
